development/data-simulator/pyproject.toml
[build-system]
requires = ["poetry>=1.0", "setuptools", "poetry_core>=1.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "simulator"
version = "0.1.0"
description = "Simulate periodic random data with anomalies"
authors = ["Domenico Di Gangi <digangidomenico@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.0"
uvicorn = "^0.23.2"
kafka-python = "^2.0.2"
numpy = "^1.25.2"
pandas = "^2.0.3"
pydantic = "^2.1.1"
tenacity = "^8.3.0"
pydantic-settings = "^2.4.0"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
simulator = "simulator.app.main:main"


